+++ common/src/main/java/io/github/mortuusars/envelope/integration/jei/EnvelopeJeiPlugin.java
@@ -5,13 +5,17 @@ import io.github.mortuusars.envelope.client.gui.screen.PackingScreen;
 import io.github.mortuusars.envelope.client.gui.screen.PaybackTagScreen;
 import io.github.mortuusars.envelope.client.util.Minecrft;
 import io.github.mortuusars.envelope.integration.jei.category.MailingRecipeCategory;
+import io.github.mortuusars.envelope.integration.jei.ingredient.EntityAddressIngredientHelper;
+import io.github.mortuusars.envelope.integration.jei.ingredient.EntityAddressIngredientRenderer;
 import io.github.mortuusars.envelope.integration.jei.util.InHandRecipeTransferInfo;
 import io.github.mortuusars.envelope.integration.jei.util.PaybackTagGhostIngredientHandler;
 import io.github.mortuusars.envelope.world.inventory.PackingMenu;
 import io.github.mortuusars.envelope.world.item.component.PackageContents;
 import io.github.mortuusars.envelope.world.item.crafting.MailRecipe;
+import io.github.mortuusars.envelope.world.mail.address.type.EntityAddress;
 import mezz.jei.api.IModPlugin;
 import mezz.jei.api.JeiPlugin;
+import mezz.jei.api.ingredients.IIngredientType;
 import mezz.jei.api.registration.*;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
@@ -24,6 +28,13 @@ import java.util.List;
 public class EnvelopeJeiPlugin implements IModPlugin {
     private static final ResourceLocation ID = Envelope.resource("jei_plugin");
 
+    public static final IIngredientType<EntityAddress> ENTITY_ADDRESS_INGREDIENT = new IIngredientType<>() {
+        @Override
+        public @NotNull Class<? extends EntityAddress> getIngredientClass() {
+            return EntityAddress.class;
+        }
+    };
+
     @Override
     public @NotNull ResourceLocation getPluginUid() {
         return ID;
@@ -32,6 +43,22 @@ public class EnvelopeJeiPlugin implements IModPlugin {
     @Override
     public void registerCategories(IRecipeCategoryRegistration registration) {
         registration.addRecipeCategories(new MailingRecipeCategory(registration.getJeiHelpers()));
+    }
+
+    @Override
+    public void registerIngredients(IModIngredientRegistration registration) {
+        List<EntityAddress> addressesWithRecipes = Minecrft.level().getRecipeManager().getAllRecipesFor(Envelope.RecipeTypes.MAILING.get())
+              .stream()
+              .map(recipe -> recipe.value().getEntityAddress())
+              .distinct()
+              .filter(address -> !address.getEntityHolder().is(Envelope.Tags.MailEntities.HIDDEN))
+              .toList();
+
+        registration.register(ENTITY_ADDRESS_INGREDIENT,
+              addressesWithRecipes,
+              new EntityAddressIngredientHelper(),
+              new EntityAddressIngredientRenderer(),
+              EntityAddress.CODEC.codec());
     }
 
     @Override